Save avant de tout PETER

+++ srcs/pipex/here_doc.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   here_doc.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ggoy <ggoy@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: roespici <roespici@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/07 16:56:05 by roespici          #+#    #+#             */
-/*   Updated: 2024/09/11 15:20:50 by ggoy             ###   ########.fr       */
+/*   Updated: 2024/09/12 09:58:25 by roespici         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,8 +28,7 @@ static void	fill_here_doc(t_pipex *pipex)
 			free(line);
 			break ;
 		}
-		if (pipex->outfile_open)
-			ft_putstr_fd(line, pipex->infile);
+		ft_putstr_fd(line, pipex->infile);
 		free(line);
 	}
 	close(pipex->infile);
@@ -37,31 +36,69 @@ static void	fill_here_doc(t_pipex *pipex)
 
 static void	exec_here_doc(t_pipex *pipex)
 {
-	pipex->infile = open("here_doc.tmp", O_RDONLY);
-	if (pipex->cmd->cmd)
-		execute_pipes(pipex);
+	if (pipex->cmd->redir->next)
+	{
+		find_last_redir(pipex->cmd, IN);
+		open_infile(pipex);
+	}
+	else
+		pipex->infile = open("here_doc.tmp", O_RDONLY);
+
+	execute_pipes(pipex);
 }
 
-static void	close_here_doc(t_pipex *pipex)
+t_lexer	*find_last_redir(t_cmd *cmd, int token)
 {
-	close_pipes(pipex);
-	unlink("here_doc.tmp");
+	t_lexer	*last;
+	t_lexer	*redir;
+
+	last = NULL;
+	redir = cmd->redir;
+	while (redir)
+	{
+		if (token == OUT || token == APPEND)
+		{
+			if (redir->token == OUT || redir->token == APPEND)
+				last = redir;
+		}
+		else if (redir->token == token)
+			last = redir;
+		redir = redir->next;
+	}
+	printf("last = %s\n", last->element);
+	return (last);
 }
 
 void	here_doc(t_pipex *pipex)
 {
-	if (pipex->cmd->redir->element)
-		pipex->limiter = ft_strdup(pipex->cmd->redir->element);
-	open_outfile(pipex);
+	t_lexer	*last;
+
+	last = find_last_redir(pipex->cmd, HEREDOC);
+	last = find_last_redir(pipex->cmd, OUT);
+	last = find_last_redir(pipex->cmd, APPEND);
+	last = find_last_redir(pipex->cmd, IN);
+	if (!last)
+		return ;
+	pipex->limiter = ft_strdup(pipex->cmd->redir->element);
 	fill_here_doc(pipex);
-	if (!pipex->outfile_open)
+	if (pipex->cmd->redir->next && pipex->cmd->redir != last)
 	{
-		printf("bash: %s: Permission denied\n", pipex->cmd->redir->element);
-		unlink("here_doc.tmp");
+		pipex->cmd->redir = pipex->cmd->redir->next;
 		free(pipex->limiter);
+		unlink("here_doc.tmp");
+		open_infile(pipex);
+	}
+	if (open_outfile(pipex) == FAILURE)
+	{
+		printf("bash: %s: Permission denied\n", last->element);
+		free(pipex->limiter);
+		unlink("here_doc.tmp");
 		return ;
 	}
-	exec_here_doc(pipex);
-	close_here_doc(pipex);
-	free(pipex->limiter);
+	if (pipex->cmd->cmd && pipex->cmd->redir == last)
+	{
+		exec_here_doc(pipex);
+		close_pipes(pipex);
+		unlink("here_doc.tmp");
+	}
 }
